+++ Framework/Router.php
@@ -38,7 +38,7 @@ class Router {
         try {
             $controller = $this->createControllerImp($module, $controller_name, false, $request);
         } catch (PfmRoutingException $e) {
-            $this->logger->error('no controller found, redirect to homepage', [
+            $this->logger->warning('no controller found, redirect to homepage', [
                 'url' => $request->getParameter('path'),
                 'controller' => $controller_name,
                 'module' => $module
@@ -51,8 +51,8 @@ class Router {
     }
 
     private function route($request) {
-
         $modulesNames = Configuration::get("modules");
+        $modulesNames = is_array($modulesNames) ? $modulesNames : [$modulesNames];
         foreach ($modulesNames as $moduleName) {
             // get the routing class
             $routingClassUrl = "Modules/" . $moduleName . "/" . ucfirst($moduleName) . "Routing.php";
@@ -84,7 +84,6 @@ class Router {
      * Examine a request and run the dedicated action
      */
     public function routerRequest() {
-
         try {
             // Merge parameters GET and POST
             $params = array();
@@ -105,7 +104,7 @@ class Router {
 
                 $urlInfo = $this->getUrlData($request);
                 if(!$urlInfo['pathInfo']) {
-                    $this->logger->error('no route found, redirect to homepage', [
+                    $this->logger->warning('no route found, redirect to homepage', [
                         'url' => $request->getParameter('path'),
                     ]);
                     $this->call('core/coretiles/index', [], $request);
@@ -123,7 +122,7 @@ class Router {
                 //$controller->runAction($urlInfo["pathInfo"]["module"], $action, $args);
             }
         } catch (Throwable $e) {
-            Configuration::getLogger()->error('[router] something went wrong', ['error' => $e->getMessage(), 'stack' => $e->getTraceAsString()]);
+            Configuration::getLogger()->warning('[router] something went wrong', ['error' => $e->getMessage(), 'stack' => $e->getTraceAsString()]);
             $this->manageError($e);
         }
     }
@@ -142,10 +141,9 @@ class Router {
                 ));
             }
         } else {
-            if($this->useRouterController){
+            if ($this->useRouterController) {
                 $controller->indexAction($args);
-            }
-            else{
+            } else {
                 $controller->runAction($urlInfo["pathInfo"]["module"], $action, $args);
             }
         }