+++ Modules/antibodies/Model/Prelevement.php
@@ -124,7 +124,6 @@ class Prelevement extends Model {
 
     public function delete($id_space, $id) {
         $sql = "UPDATE ac_prelevements SET deleted=1,deleted_at=NOW() WHERE id=? AND id_space?";
-        //$sql = "DELETE FROM ac_prelevements WHERE id = ?";
         $this->runRequest($sql, array($id, $id_space));
     }