+++ Modules/antibodies/Model/Tissus.php
@@ -181,7 +181,6 @@ class Tissus extends Model {
 
     public function delete($id_space, $id) {
         $sql = "UPDATE ac_j_tissu_anticorps SET deleted=1,deleted_at=NOW() WHERE id=? AND id_space=?";
-        //$sql = "DELETE FROM ac_j_tissu_anticorps WHERE id = ?";
         $this->runRequest($sql, array($id, $id_space));
     }