+++ Modules/antibodies/View/Antibodieslist/indexAction_1.php
@@ -355,7 +355,7 @@
                             if ($tissus[$i]['ref_protocol'] == "0") {
                                 $val .= "<p>Manuel</p>";
                             } else {
-                                $val .= "<p><a href=\"protocols/protoref/" . $anticorps ['id'] . "\">"
+                                $val .= "<p><a href=\"protocols/protoref/" . $anticorps ['id'] . "/" . $id_space . "\">"
                                         . $tissus[$i]['ref_protocol'] . "</a></p>";
                             }
                         }