+++ Modules/booking/Model/BkAuthorization.php
@@ -647,7 +647,6 @@ class BkAuthorization extends Model {
      */
     public function delete($id_space, $id) {
         $sql = "UPDATE bk_authorization set deleted=1,deleted_at=NOW() WHERE id = ? AND id_space=?";
-        //$sql = "DELETE FROM bk_authorization WHERE id = ? AND id_space=?";
         $this->runRequest($sql, array($id, $id_space));
     }