+++ Modules/bulletjournal/Model/BjNote.php
@@ -153,7 +153,7 @@ class BjNote extends Model {
      */
     public function delete($id_space, $id) {
         $sql = "DELETE FROM bj_notes WHERE id = ? AND id_space=?";
-        $this->runRequest($sql, array($id));
+        $this->runRequest($sql, array($id, $id_space));
     }
 
 }