+++ Modules/resources/Model/ReVisa.php
@@ -270,8 +270,7 @@ class ReVisa extends Model {
      */
     public function delete($id_space, $id) {
         $sql = "UPDATE re_visas SET deleted=1,deleted_at=NOW() WHERE id=? AND id_space=?";
-        // $sql = "DELETE FROM re_visas WHERE id = ? AND id_space=?";
-        $this->runRequest($sql, array($id));
+        $this->runRequest($sql, array($id, $id_space));
     }
 
     public function getAllInstructors($id_space) {